requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
celery>=5.3,<6.0
redis>=5.0,<6.0
channels>=4.0,<5.0
elastic-apm>=6.3,<7.0
python-logstash>=0.4.6,<0.5
networkx>=3.2,<4.0
